fix the test
